explicitly convert to float

+++ data-logger/python/analyze_evaluation_run.py
@@ -54,10 +54,10 @@ def analyzedatasets(main_dir,subdirs,prefix, results_dir="results", plot=False):
     resultsdict["mean_time_between_failures"] = mtbf.tolist()
     resultsdict["mean_distance_between_failures"] = mdbf.tolist()
 
-    resultsdict["grandmean_failure_scores"] = np.mean(mean_failure_scores)
-    resultsdict["grandmean_num_failures"] = np.mean(num_failures)
-    resultsdict["grandmean_time_between_failures"] = np.mean(mtbf)
-    resultsdict["grandmean_distance_between_failures"] = np.mean(mdbf)
+    resultsdict["grandmean_failure_scores"] = float(np.mean(mean_failure_scores)))
+    resultsdict["grandmean_num_failures"] = float(np.mean(num_failures))
+    resultsdict["grandmean_time_between_failures"] = float(np.mean(mtbf))
+    resultsdict["grandmean_distance_between_failures"] = float(np.mean(mdbf))
     print(resultsdict)
     with open(results_fp,'w') as f:
         yaml.dump(resultsdict,f,Dumper=yaml.SafeDumper)
